Module kth::quickselect

Efficiently locate the k-th element of an array. The k-th element is the element that would be at index k if the array were sorted.

Finding the k-th element is generally faster than sorting the array.

Algorithms here are largely based on [SEA 207].

SEA 207

Another paper by Andrei Alexandrescu on this topic: https://arxiv.org/pdf/1606.00484v1.pdf

Functions

hoare_partition
Partition an array (move all elements greater than a given element to one side, and all elements larger than the same element to the other).

median_of_medians
Find the median of medians (recursively).

quickselect
Find the k-th smallest element in a, place it at a[k] and partition a around that element (all smaller than the element to the left, all larger to the right).

repeated_step3
median-of-medians on groups of 3 elements

sort3
Sort 3 elements. This is equivalent to a theoretical partition3.

sort5
Sort 5 elements in-place