Function kth::quickselect::quickselect

pub fn quickselect<T: Ord, P: Fn(&mut [T]) -> usize>(
    partition: P, 
    a: &mut [T], 
    k: usize
)

Find the k-th smallest element in a, place it at a[k] and partition a around that element (all smaller than the element to the left, all larger to the right).

partition is a function that chooses & returns a pivot index (p) and partitions a around a[p].

a is the input array

k is the order of the desired element